+++ src/bookshop/bean/CartDataBean.java
@@ -7,7 +7,7 @@ public class CartDataBean {
 	private int book_id; //구매된 책의 아이디
 	private String book_title; //구매된 책제목
 	private int buy_price; //판매가
-	private int buy_count;// 판매 수량
+	private byte buy_count;// 판매 수량
 	private String book_image; //책 이미지
 	
 	
@@ -41,10 +41,11 @@ public class CartDataBean {
 	public void setBuy_price(int buy_price) {
 		this.buy_price = buy_price;
 	}
-	public int getBuy_count() {
+	
+	public byte getBuy_count() {
 		return buy_count;
 	}
-	public void setBuy_count(int buy_count) {
+	public void setBuy_count(byte buy_count) {
 		this.buy_count = buy_count;
 	}
 	public String getBook_image() {
